refactor(RouterTree): rename mod from handler_selector
