+++ src/lib.rs
@@ -2,7 +2,6 @@
 #![feature(return_position_impl_trait_in_trait)]
 pub mod deserializer;
 pub mod error;
-pub mod handle_selector;
 pub mod handler;
 pub mod middleware;
 pub mod request;
@@ -13,3 +12,5 @@ pub mod runner_input;
 pub mod runner_output;
 pub mod serializer;
 pub mod server;
+pub mod tree;
+
